requirements.txt
pgnparser
